requirements.txt
django
djangorestframework
nltk
numpy
pymongo
pandas
dnspython
python-dotenv